fix: implemented required changes
