+++ app/globals.css
@@ -25,3 +25,16 @@ body {
     text-wrap: balance;
   }
 } */
+
+
+@layer utilities {
+
+  .no-spin::-webkit-outer-spin-button,
+  .no-spin::-webkit-inner-spin-button {
+    @apply appearance-none m-0;
+  }
+
+  .no-spin[type='number'] {
+    @apply appearance-none;
+  }
+}
